requirements.txt
# hntldr dependencies
python-telegram-bot>=22.0
anthropic>=0.39.0
openai>=1.0.0
aiohttp>=3.9.0
trafilatura>=1.8.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
